chore: :hammer: change orderby createdAt,

+++ src/server/api/routers/desa.ts
@@ -1,6 +1,5 @@
 import { TRPCError } from "@trpc/server";
 import { count, eq, ilike } from "drizzle-orm";
-import z from "zod";
 import {
 	formatResponse,
 	formatResponseArray,
@@ -8,7 +7,7 @@ import {
 } from "@/helper/response.helper";
 import { idBase } from "@/types";
 import { desaCreateSchema, desaFilter, desaUpdateSchema } from "@/types/desa";
-import { desa, kelompok } from "../../db/schema";
+import { desa, kelompok, log } from "../../db/schema";
 import { createTRPCRouter, protectedProcedure } from "../trpc";
 
 export const desaRouter = createTRPCRouter({
@@ -33,6 +32,7 @@ export const desaRouter = createTRPCRouter({
 				limit,
 				offset: page * limit,
 				where: ilike(desa.nama, `%${input.q}%`),
+				orderBy: (desa, { desc }) => [desc(desa.createdAt)],
 			});
 
 			const [total] = await ctx.db.select({ count: count() }).from(desa);
@@ -40,25 +40,11 @@ export const desaRouter = createTRPCRouter({
 			const totalCount = total?.count ?? 0;
 			const totalPages = Math.ceil(totalCount / limit);
 
-			// relation check
-			const items = await Promise.all(
-				data.map(async (d) => {
-					const relatedKelompok = await ctx.db.query.kelompok.findFirst({
-						where: eq(kelompok.desaId, d.id),
-					});
-
-					return {
-						...d,
-						isDeletable: !relatedKelompok,
-					};
-				}),
-			);
-
 			return formatResponsePagination(
 				true,
 				"Berhasil mendapatkan data Desa",
 				{
-					items,
+					items: data,
 					meta: { total: totalCount, page, limit, totalPages },
 				},
 				null,
@@ -67,7 +53,7 @@ export const desaRouter = createTRPCRouter({
 
 	getOneDesa: protectedProcedure.input(idBase).query(async ({ ctx, input }) => {
 		const data = await ctx.db.query.desa.findFirst({
-			where: eq(desa.id, Number(input.id)),
+			where: eq(desa.id, input.id),
 		});
 
 		if (!data) {
@@ -86,14 +72,20 @@ export const desaRouter = createTRPCRouter({
 			const existingDesa = await ctx.db.query.desa.findFirst({
 				where: eq(desa.nama, input.nama),
 			});
-
 			if (existingDesa) {
 				throw new TRPCError({
 					code: "BAD_REQUEST",
 					message: "Data Desa sudah ada",
 				});
 			}
+
 			const data = await ctx.db.insert(desa).values(input);
+
+			await ctx.db.insert(log).values({
+				event: "Create",
+				description: `Menambahkan Data Desa ${input.nama}`,
+				userId: ctx.session.user.email,
+			});
 
 			return formatResponse(true, "Berhasil menambahkan data Desa", data, null);
 		}),
@@ -101,6 +93,16 @@ export const desaRouter = createTRPCRouter({
 	updateDesa: protectedProcedure
 		.input(desaUpdateSchema)
 		.mutation(async ({ ctx, input }) => {
+			const existingDesa = await ctx.db.query.desa.findFirst({
+				where: eq(desa.nama, input.nama),
+			});
+			if (existingDesa) {
+				throw new TRPCError({
+					code: "BAD_REQUEST",
+					message: "Data Desa sudah ada",
+				});
+			}
+
 			if (!input.id) {
 				throw new TRPCError({
 					code: "BAD_REQUEST",
@@ -113,19 +115,29 @@ export const desaRouter = createTRPCRouter({
 				.set(input)
 				.where(eq(desa.id, input.id));
 
+			await ctx.db.insert(log).values({
+				event: "Update",
+				description: `Mengubah Data Desa ${input.nama}`,
+				userId: ctx.session.user.email,
+			});
+
 			return formatResponse(true, "Berhasil mengubah data Desa", data, null);
 		}),
 
 	deleteDesa: protectedProcedure
-		.input(
-			z.object({
-				id: z
-					.int()
-					.min(1, "ID tidak boleh kosong")
-					.max(32767, "ID maksimal 32767"),
-			}),
-		)
+		.input(idBase)
 		.mutation(async ({ ctx, input }) => {
+			const relatedKelompok = await ctx.db.query.kelompok.findFirst({
+				where: eq(kelompok.desaId, input.id),
+			});
+
+			if (relatedKelompok) {
+				throw new TRPCError({
+					code: "BAD_REQUEST",
+					message: "Data Kelompok terkait ditemukan",
+				});
+			}
+
 			const data = await ctx.db.delete(desa).where(eq(desa.id, input.id));
 
 			return formatResponse(true, "Berhasil menghapus data Desa", data, null);
